requirements.txt
pillow
ollama
ultralytics
insightface
onnxruntime